Menu classes uses unsingned int type instead of size_t, because they need and int instead of long. #154

+++ src/game/ui/screen/game_ending_screen.cpp
@@ -37,7 +37,7 @@ void game_ending_screen::initialize()
     add_menu_item(asset_path::main_menu_image);
     add_menu_item(asset_path::exit_image);
 
-    select_item(static_cast<size_t>(menu_option::restart));
+    select_item(static_cast<unsigned int>(menu_option::restart));
 }
 
 void game_ending_screen::update()
